requirements.txt
opencv-python==4.11.0.86
Flask==3.1.0
Flask-SocketIO==5.5.1
numpy==2.2.3
scipy==1.15.2
cvxopt==1.3.2
matplotlib==3.10.1
pillow==11.1.0
flask-sqlalchemy==3.1.1
pymysql==1.1.1
pyjwt==2.10.1
Werkzeug~=3.1.3
python-dotenv==1.0.1
